Add Getting Started section
Use information from previous launch locally to create a Getting Started guide. Notably, the getting started documentation explicitly points to the demo documentation to load data.

Remove project structure section to avoid out-dated information based on new intended deployment structure. It is a reasonable expectation that a new user does not need to see this on his first visit of the repository. They can open the folder structure or visit the other documentation for more information.